+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2023-09-28
+
+### Added
+
+- Add an option to toggle the vertex displacement in PBR refinement.
+- Add Input Image operator.
+
+### Changed
+
+- Change Cylinder and Cube in PBR refinement.
+- No need to shift + click to rotate the preview in PBR refinement.
+
 ## [0.2.0] - 2023-09-20
 
 ### Added